+++ src/mech/ct_log.h
@@ -44,35 +44,35 @@ extern "C" {
 		}                                                                                                   \
 	} while (0)
 
-#define CTLogger_HandleHex(__flag, __type, __buf, __len)                                    \
-	do {                                                                                    \
-		if ((__len) > 0 && ct_log_is_enable(__type, CTLog_Level##__flag)) {                 \
-			ct_threadcache_t *__cache     = ct_threadcache_get();                           \
-			char             *__buffer    = ct_threadcache_get_buffer(__cache);             \
-			char             *__dst       = __buffer;                                       \
-			const uint8_t    *__src       = (const uint8_t *)(__buf);                       \
-			size_t            __available = 1024;                                           \
-			const char       *__hex_table = "0123456789ABCDEF";                             \
-			for (size_t __i = 0; __i < (size_t)(__len); __i++) {                            \
-				if (__available < 3) {                                                      \
-					ct_log_handle(__type, CTLog_Level##__flag, __buffer, __dst - __buffer); \
-					__dst       = __buffer;                                                 \
-					__available = 1024;                                                     \
-				}                                                                           \
-				const uint8_t byte = __src[__i];                                            \
-				*__dst++           = __hex_table[byte >> 4];                                \
-				*__dst++           = __hex_table[byte & 0x0F];                              \
-				if (__i != (size_t)(__len) - 1) {                                           \
-					*__dst++ = ' ';                                                         \
-					__available -= 3;                                                       \
-				} else {                                                                    \
-					__available -= 2;                                                       \
-				}                                                                           \
-			}                                                                               \
-			if (__dst > __buffer) {                                                         \
-				ct_log_handle(__type, CTLog_Level##__flag, __buffer, __dst - __buffer);     \
-			}                                                                               \
-		}                                                                                   \
+#define CTLogger_HandleHex(__flag, __type, __buf, __len)                                              \
+	do {                                                                                              \
+		if ((__len) > 0 && ct_log_is_enable(__type, CTLog_Level##__flag)) {                           \
+			ct_threadcache_t *__cache     = ct_threadcache_get();                                     \
+			char             *__buffer    = ct_threadcache_get_buffer(__cache);                       \
+			char             *__dst       = __buffer;                                                 \
+			const uint8_t    *__src       = (const uint8_t *)(__buf);                                 \
+			size_t            __available = 1024;                                                     \
+			const char       *__hex_table = "0123456789ABCDEF";                                       \
+			for (size_t __i = 0; __i < (size_t)(__len); __i++) {                                      \
+				if (__available < 3) {                                                                \
+					ct_log_handle(__type, CTLog_Level##__flag, __buffer, (size_t)(__dst - __buffer)); \
+					__dst       = __buffer;                                                           \
+					__available = 1024;                                                               \
+				}                                                                                     \
+				const uint8_t byte = __src[__i];                                                      \
+				*__dst++           = __hex_table[byte >> 4];                                          \
+				*__dst++           = __hex_table[byte & 0x0F];                                        \
+				if (__i != (size_t)(__len) - 1) {                                                     \
+					*__dst++ = ' ';                                                                   \
+					__available -= 3;                                                                 \
+				} else {                                                                              \
+					__available -= 2;                                                                 \
+				}                                                                                     \
+			}                                                                                         \
+			if (__dst > __buffer) {                                                                   \
+				ct_log_handle(__type, CTLog_Level##__flag, __buffer, (size_t)(__dst - __buffer));     \
+			}                                                                                         \
+		}                                                                                             \
 	} while (0)
 
 #ifdef __cplusplus
